chess-api/requirements.txt
ultralytics==8.3.0
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
python-dotenv==1.0.1